# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Notifications microservice for FiuFit's backend.",
        "title": "Notifications Microservice API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "Returns a 200 code and JSON with a message.",
                "summary": "Check the service's status.",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns a 200 code and JSON with the date the service started running and a description.",
                "summary": "Get the service's health status.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.HealthResponse"
                        }
                    }
                }
            }
        },
        "/logs": {
            "get": {
                "description": "Get the service's logs.",
                "produces": [
                    "text/plain"
                ],
                "summary": "Get the service's logs.",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/new-follower": {
            "post": {
                "description": "Send new follower notification to the given followed user.",
                "summary": "Send new follower notification to the given followed user.",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Get the status of the service.",
                "produces": [
                    "application/json"
                ],
                "summary": "Get the service's blocked status.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Decoded payload of the admin token",
                        "name": "user_info",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ChangeStatusInput"
                        }
                    }
                }
            },
            "post": {
                "description": "Changes the status of the service.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Change the service's blocked status.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Decoded payload of the admin token",
                        "name": "user_info",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Blocked status of the service",
                        "name": "rule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ChangeStatusInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.ChangeStatusInput"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ChangeStatusInput": {
            "type": "object",
            "properties": {
                "blocked": {
                    "type": "boolean"
                }
            }
        },
        "controller.HealthResponse": {
            "type": "object",
            "properties": {
                "creation_date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "last_response": {
                    "type": "string"
                }
            }
        }
    }
}